+++ dashboard/crossview.R
@@ -21,7 +21,7 @@ sta_avg <- metrosp::station_averages |>
   filter(date >= DATA_START) |>
   mutate(line_number = as.character(line_number)) |>
   filter(line_number %in% LINES) |>
-  select(date, line_number, station_name, value = value, year)
+  select(date, line_number, station_id, station_name, value = value, year)
 
 ## Station demand summary (last 12 months) ----
 latest_date <- max(sta_avg$date, na.rm = TRUE)
@@ -41,10 +41,10 @@ coords <- sf::st_coordinates(sf_stations)
 stations_geo <- sf_stations |>
   sf::st_drop_geometry() |>
   mutate(lng = coords[, 1], lat = coords[, 2]) |>
-  select(station_name, line_number, lng, lat)
+  select(station_id, station_name, line_number, lng, lat)
 
 station_map_data <- sta_summary |>
-  inner_join(stations_geo, by = c("line_number", "station_name"))
+  inner_join(stations_geo, by = c("line_number", "station_id"), suffix = c("", "_geo"))
 
 ## Line geometries ----
 sf_lines <- tryCatch(